backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
pytest==7.4.4
httpx==0.26.0
